thing

+++ src/huf.c
@@ -231,8 +231,8 @@ send_block( /* void */ )
 /* lh4, 5, 6, 7 */
 void
 output_st1(c, p)
-    unsigned short  c;
-    unsigned short  p;
+    int  c;
+    int  p;
 {
     static unsigned short cpos;
